package.json: use ${workspaceRoot}
At the moment we magically append the workspace root to some properties of the launch.json on the vscode side.
This will go away in the future since it is clunky. Thus we want to educate users to use the ${workspaceRoot} variable for all paths, since we will no longer support relative paths in the future.