Make HoverProvider tolerate missing locations
GoHoverProvider was throwing a TypeError trying to read a property
from null definition locations. It may be more appropriate to try
and prevent the hover provider from entering its handler function if
definitionLocation finds no match, but rejecting the Promise didn't
see to do the trick. This patch makes the HoverProvider call a no-op
when no definition is found..