Support github.com/josharian/impl. Fixes #259.
Does some basic input validation to ensure the required information is
passed.

The regex included makes the receiver name optional, the intention there
is that if the receiver name isn't included, that the type the cursor is
in should be used as the receiver for the interface.

However, I think that the first step of just having the ability to use
goimpl was a important first step.