Run on-save tools for entire workspace (#1023)
* Vet and lint entire workspace

Sets the current working directory to the workspace root when linting
and vetting. Also appends the './...' argument which causes the tools to
traverse the workspace recursively. Other tools are not affected.

Additionally, this will utilize 'go vet' instead of 'go tool vet'.

* Run build (on save) on entire workspace

Sets working directory for 'go build' command to workspace root. Also
adds './...', so the entire workspace is traversed.

* Allow users to change the scope for on-save tools

This modifies settings 'buildOnSave', 'lintOnSave' and 'vetOnSave'. The
new options are 'workspace', 'package' or 'off' which will cause the
respective on-save tools to run in the determined scope.

Old values will still work as they did before.

* Build on-save shall not trigger test of entire workspace

'buildOnSave' compiles the test for the current package if the saved
file is a test unit, no matter if it is set to 'workspace' or 'package'.

* Compile tests for entire workspace if settingss require

If 'buildOnSave' is set to 'workspace', the check function will traverse
the workspace in order to find sub-packages that can be tested and
subsequently compiles those tests.

* Check tests for entire workspace when settings require it

Utilize 'go test' to check all tests in the workspace without actually
running them.

* Prevent running both 'go build' and 'go test' on save

Utilize 'go test -run=^$ ./...' when building entire workspace, 'go build'
for current package and 'go test -c' for the current package when the
current file is a test file.