Add command to fetch import on current line (#1222)
* Add command to fetch import on current line

This will allow users to easily fetch a dependency without having to
take their attention away from the code they're editing.

Microsoft/vscode-go#1189

* Rename goGetImports to goGetImport

The plugin only fetches one import so having the file name plural is
misleading to anybody reading the code.

Microsoft/vscode-go#1189

* Refactor getImportPath

This consolidates the functionality found in two commands into one place
so that if it ever needs to be changed or removed it can be done more
easily.

* Fix lint issues

* Make command title and description more intuitive

Microsoft/vscode-go#1189

* Make name `goGetPackage` consistent

Anybody else who comes along later and reads this is gonna have an
easier time if they only have to keep one name in their head vs. having
to keep two (goGetImport and goGetPackage).

Microsoft/vscode-go#1189

* Exit early if there's no import path to get

If there's no import path to pass to `go get` then there's no point in
spinning up a child process and waiting for it to finish to error out.

This also fixes a bug in `getImportPath` that would cause it to return
invalid input paths for things like comments and function definitions.

Microsoft/vscode-go#1189

* Fallback to selectedText for browsePackage feature

* Offload check for package to go get

`go get` won't download a package if it's already in `$GOPATH/src`.
There's no point in checking for the package to exist and then
downloading with `-u` because `go get -u` and `go get` are effectively
the same when a package hasn't been downloaded yet.

Microsoft/vscode-go#1189

* Use output pane for go get output

`go get -v` gives users more information on what's going on with the
command. This way a user can tell if the package was even downloaded,
because by default `go get` won't output anything.

`go get -v` will output nothing if the package isn't downloaded because
it already exists. To get around this, an information message is shown
saying that the package wasn't downloaded. This keeps the user informed
because they will be expecting feedback from the command.

In the event of an error (like if there's no internet connection), the
error gets shown to the output pane so the user knows what happened.

Microsoft/vscode-go#1189

* Cleanup