Display nested content in the variables pane
If a variable contains children but they have not been loaded issue an
expression command to delve to fetch them, then update the variable's
children array with the result. This required refactoring the
evaluateRequest method into two parts so that delve expressions can be
evaluated synchronously.
Removed unreferenced param in convertDebugVariableToProtocolVariable.