Changed swapped $2 and $0 position for anonymous go function snippet. (#2354)
For anonymous go function (the snippet with the prefix "go") I think it will be more convenient if we end snippets in the body instead of the variable field. The function body is the largest of these 3 placeholders and will be in need of completion features more than any other. Therefore, I think, it is reasonable if we give it the $0 place.