build/all.bash: script to enable running tests locally and reliably
This script provides a command to build and run tests in a docker
container where all tools and source code used by tests are installed
and all environment variables tests depend on (e.g. GOPATH) are set
properly.

   Usage: build/all.bash testlocal

The container uses the latest official release of Go and node.
I hoped this could help setting up the environment to reproduce the
current CI test run failure but the test succeeded - that indicates the
problem may be in the environment used in the travis.