goLanguageServer.ts: allow Go Nightly to use pre-release versions.
Also add isForNightly that is true if the extension is for nightly

The extension name is one of the best way to check the extension's mode,
but I couldn't find a way to dynamically determine the
current extensionId through vscode api. For now I use the already existing
constant that happens to be in the telemetry.ts.
The extensionId is useful for purposes other than telemetry, so
we'd better consider refactoring in the future.

NIGHTLY ONLY

Change-Id: I00377edbca233d1951052d200c730001493dc3cd