extend installTools to choose the version
That allows us to plumb the version picked from goLanguageServer's
latestGopls and request the installation of the specific version.
If the version is not specified in ToolAtVersion, it implies
the latest version of the tool.

Tested manually while gopls v0.3.0-preN became available.

Also, print the version in the installation result output.

Change-Id: Ic089bf15a0586c419440a85318d2c7976d1af62a