Update to latest vscode-languageclient
This reverts commit 5f15c237c71e527d983d29d351597838868c4b1b.

This effectively brings in the commit 3b02de0 back again.
Originally reverted because it requires vscode 1.41.x,
which was just released at the time of Go extension 0.13.0
release. Currently 1.42.x is the latest version, and we expect
another cycle of vscode release when the next version of Go
extension is out.

We need more up-to-date version of vscode-languageclient to
bring in various fixes to problems such as
https://github.com/microsoft/vscode-languageserver-node/issues/442

Considered only minor version upgrade but I couldn't find a suitable
newer version than 5.2.1
(https://www.npmjs.com/package/vscode-languageclient)