github/workflows: add CI github action workflow
Ideally I want the CI to run from Gerrit because that will be our primary change review system,
but the internal system is not ready yet.
This is a temporary solution that uses Github Action. The CI workflow kicks in for every PR and every commit on the master.

Change-Id: I97c9d20f4649c4cafd7a984c02fa84d0bc70f554
GitHub-Last-Rev: 948b555fc413ccff9d4c8e2f4564af041f73bbbc
GitHub-Pull-Request: golang/vscode-go#4
Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/222921
Reviewed-by: Rebecca Stambler <rstambler@golang.org>