Fix 'run without debugging' when using modules or no user debug configuration (#3125)
* debug: fix 'run without debugging'

Bug 1: when generating the default debug configuration
because there is no user-configured one, the extension
dropped noDebug field and caused the program to run with
debugging enabled. Fix it (goDebugConfiguration.ts)

Bug 2: when debug adapter receives noDebug request, it
runs the program with `go run` instead of invoking the
program through dlv. The `go run` will not work in modules
mode if the command runs outside the main module. Set
cwd accordingly.

Fixes microsoft/vscode-go#3121

TESTED=manually with the example attached in #3121

* Handle case when debugConfiguration doesnt exist

Co-authored-by: Ramya Achutha Rao <ramyar@microsoft.com>