test/integration: fix Build Tags checking tests
The test invoked 'check' concurrently with different sets of tags.
Underneath, the check function invokes the goBuild, which cancels
outstanding go build commands before starting another go build command.
That means, the three go build processes race and try to cancel
others. That resulted in empty diagnostics and thus, test failures.

Serialize the tests.

While we are here, also fixes the use of assert.equal which expects
the actual value as the first parameter and the expected value as
the second parameter.